requirements.txt
langchain
langchain-chroma
streamlit
langchain-groq
python-dotenv
pandas
langchain-ollama
langchain-community
langchain-pinecone
sentence-transformers